优化逻辑 + 1

+++ gk-platform/src/main/java/com/gk/platform/controller/SysDeptController.java
@@ -1,13 +1,18 @@
 package com.gk.platform.controller;
 
+import cn.hutool.core.util.ObjUtil;
 import com.gk.common.annotation.RequiresPermission;
-import com.gk.common.tools.R;
+import com.gk.common.constant.Constant;
+import com.gk.common.context.ReqContextHolder;
+import com.gk.common.exception.ErrorCode;
+import com.gk.common.model.R;
 import com.gk.common.validator.AssertUtils;
 import com.gk.platform.dto.SysDeptDTO;
 import com.gk.platform.service.SysDeptService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -27,16 +32,22 @@ public class SysDeptController {
 
     @GetMapping("list")
     @Operation(summary = "列表")
-    @RequiresPermission("sys:dept:list")
+    @PreAuthorize("hasAuthority('sys:dept:view')")
     public R<?> list() {
-        List<SysDeptDTO> list = sysDeptService.list(new HashMap<>(1));
+        List<SysDeptDTO> list = sysDeptService.list(new HashMap<>());
         return R.ok(list);
     }
 
     @PostMapping
     @Operation(summary = "保存")
-    @RequiresPermission("sys:dept:save")
+    @PreAuthorize("hasAuthority('sys:dept:add')")
     public R<?> save(@RequestBody SysDeptDTO dto) {
+        if (!ReqContextHolder.isSAdmin()) {
+            dto.setTenantId(ReqContextHolder.getTenantId());
+            if (ObjUtil.isEmpty(dto.getPid()) || dto.getPid() <= Constant.MAX_RESERVED_ID) {
+                dto.setPid(ReqContextHolder.getDeptId());
+            }
+        }
         //效验数据
         sysDeptService.save(dto);
         return R.ok();
@@ -44,9 +55,15 @@ public class SysDeptController {
 
     @PutMapping("{id}")
     @Operation(summary = "修改")
-    @RequiresPermission("sys:dept:update")
+    @PreAuthorize("hasAuthority('sys:dept:update')")
     public R<?> update(@PathVariable("id") Long id, @RequestBody SysDeptDTO dto) {
         dto.setId(id);
+        if (!ReqContextHolder.isSAdmin()) {
+            dto.setTenantId(null);
+            if (ObjUtil.isEmpty(dto.getPid()) || dto.getPid() <= Constant.MAX_RESERVED_ID) {
+                dto.setPid(ReqContextHolder.getDeptId());
+            }
+        }
         //效验数据
         sysDeptService.update(dto);
         return R.ok();
@@ -54,10 +71,13 @@ public class SysDeptController {
 
     @DeleteMapping("{id}")
     @Operation(summary = "删除")
-    @RequiresPermission("sys:dept:delete")
+    @PreAuthorize("hasAuthority('sys:dept:delete')")
     public R<?> delete(@PathVariable("id") Long id) {
         //效验数据
         AssertUtils.isNull(id, "id");
+        if(id <= 100){
+            return R.error(ErrorCode.FORBIDDEN);
+        }
         sysDeptService.delete(id);
         return R.ok();
     }
